// front-end/src/app/import-contacts-module/import-contacts/model/error-contact.model.ts
import { ErrorFieldModel } from './error-field.model';

export class ErrorContactModel {
  id: ErrorFieldModel;
  committeeId: ErrorFieldModel;
  type: ErrorFieldModel;
  name: ErrorFieldModel;
  lastName: ErrorFieldModel;
  firstName: ErrorFieldModel;
  middleName: ErrorFieldModel;
  suffix: ErrorFieldModel;
  prefix: ErrorFieldModel;
  street: ErrorFieldModel;
  street2: ErrorFieldModel;
  city: ErrorFieldModel;
  state: ErrorFieldModel;
  zip: ErrorFieldModel;
  employer: ErrorFieldModel;
  occupation: ErrorFieldModel;
  candidateId: ErrorFieldModel;
  officeSought: ErrorFieldModel;
  officeState: ErrorFieldModel;
  district: ErrorFieldModel;
  multiCandidateCmteStatus: ErrorFieldModel;

  constructor() { }
}